[EditModel] Re-introduce the converters for fields
Run the field converters when they are available on the
fieldconfig. This way we can joggle field types where there
are inconsistencies between the schema and the UI.
In the case of decimals in output this is a dropdown in the UI
however it is represented as an integer in the schema.

Closes #114