Implement reload button for related types
A add and reload button have been added for types
that are also managed by the maintenance app.
The add button will open a new tab that the user can use
to add a new type of the value to be selected. In the previous
screen the refresh button can be used to reload the options
after the value has been added.

Implements #133 partially, does not auto select the newly created
value.

References #133