Use field value converters for program indic
The program indicator values were not run through the
converters that can be set on the fieldConfig.
The decimals field is one of the fields that was effected
by this, this resulted into the correct value in the dropdown
not being selected and the value being sent to the server as a
string instead of a number.

DHIS2-1914