Use the central store for snackbar (#357)
Refactor snackbar to use the central application store instead of its own separate store. This change revealed a flaw that led to circular dependencies (causing the tests to break even though the app worked in the browser). What was happening was that the event-program and program-indicator reducers imported a module (e.g. event-program-steps) which then imported UI components. So reducers were basically importing components. And all hell broke loose.

To fix the circular dependencies, event-program-steps and program-indicator-steps were refactored to only hold information about the steps, and not the actual components. The setting up of the components was moved to the respective *StepperContent files, which is the only place the components were actually used.

Summary of changed files:

/EditModel/* : refactoring to handle circular dependencies

/Snackbar/* : switch to central store