DHIS2-2683 Remove callbacks from the `onEnter` handlers (#381)
Due to how HashHistory is implemented when it changes URL hashes
in v3 of react-router multiple events fires when we use
`HashHistory.push(url)`. This in turn causes the `onEnter` handler
in react-router to run twice.

The XHR request itself is cached so no extra data is transmitted from
the server but another consequence is that the components render twice
which has performance implications.

For more information see the [JIRA
issue](https://jira.dhis2.org/browse/DHIS2-2683).