fix(optionset): reset changedOriginalFieldValues when model changes (#611)
* fix(optionset): reset changedOriginalFieldValues when model changes

* the why comment

* Use resultselctor to create optionForm

* Revert changed model values when clicking cancel (#599)

* fix: option-set
Previously it was testing if the value of a given property key was undefined. However, this can produce incorrect behavior, because undefined can be a valid value of a d2 model, i.e. when a style object has not been set yet. So when doing set-color+cancel, and then again set-color+cancel, on the second run it would set the cancelled color as the value on `changedOriginalFieldValues`. Now we are testing if the `changedOriginalFieldValues` has the given property using the built in `hasOwnProperty`. This cannot give false positives as `var test = { a: undefined }; test.hasOwnProperty('a')` will return `true`