fix: support categoryMappings for program and programIndicator
categoryMappings are an embeddedObject, and we need this patch for model.save()
and getOwnedPropertyJSON to include the full categoryMappings object.
Similarly, we need categoryMappingIds to post the verbatim array for programIndicator.

Adding this as a patch, because the change is in d2. Which is deprecated,
as well as the app being stuck on an old version, which is not published on CI